requirements.txt
openai==0.27.2
python-dotenv~=1.0.0
sqlalchemy==2.0.6